♻️ Fix WorkPage Entire page deopted into client-side rendering
