+++ app/works/WorkList.module.scss
@@ -1,0 +1,14 @@
+@import 'styles/variables';
+
+.cards {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-gap: 12px;
+
+  @media screen and (max-width: 750px) {
+    grid-template-columns: repeat(2, 1fr);
+  }
+  @media screen and (max-width: 504px) {
+    grid-template-columns: 1fr;
+  }
+}
