+++ app/works/page.module.scss
@@ -1,4 +1,5 @@
 @import 'styles/variables';
+
 .title {
   font-weight: 700;
   margin-top: 4px;
@@ -7,19 +8,6 @@
     font-size: 1.1rem;
     font-weight: 500;
     margin: 0px;
-  }
-}
-
-.cards {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-gap: 12px;
-
-  @media screen and (max-width: 750px) {
-    grid-template-columns: repeat(2, 1fr);
-  }
-  @media screen and (max-width: 504px) {
-    grid-template-columns: 1fr;
   }
 }
